requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.12.3
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
dataclasses-json==0.6.7
faiss-cpu==1.13.2
fastapi==0.136.1
filelock==3.29.0
frozenlist==1.8.0
fsspec==2026.4.0
google-generativeai==0.8.5
greenlet==3.5.0
h11==0.16.0
hf-xet==1.5.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.13.0
idna==3.13
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
langchain-classic==1.0.6
langchain-community==0.4.1
langchain-core==1.3.3
langchain-protocol==0.0.15
langchain-text-splitters==1.1.2
langsmith==0.8.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.4
orjson==3.11.8
packaging==26.2
propcache==0.4.1
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
pypdf==6.10.2
PyPDF2==3.0.1
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
setuptools==81.0.0
shellingham==1.5.4
SQLAlchemy==2.0.49
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.8.0
typer==0.25.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.46.0
watchfiles==1.1.1
websockets==16.0
xxhash==3.7.0
yarl==1.23.0
zstandard==0.25.0
